+++ media/smart-card/src/__tests__/visual-regression/hover-card.test.tsx
@@ -17,15 +17,37 @@ describe('Hover Card', () => {
       '[data-testid="smart-element-icon-icon--wrapper"]',
     );
 
-    const image = await takeSnapshot(page, height);
-    expect(image).toMatchProdImageSnapshot();
+    return page;
   };
 
   it('should open below trigger component when there is room below in viewport', async () => {
-    await renderHoverCard(650);
+    const height = 650;
+
+    const page = await renderHoverCard(height);
+
+    const image = await takeSnapshot(page, height);
+    expect(image).toMatchProdImageSnapshot();
   });
 
   it('should open on top of trigger component when there is no room below and there is room above in viewport', async () => {
-    await renderHoverCard(350);
+    const height = 350;
+
+    const page = await renderHoverCard(height);
+
+    const image = await takeSnapshot(page, height);
+    expect(image).toMatchProdImageSnapshot();
+  });
+
+  it('should open preview modal', async () => {
+    const page = await renderHoverCard();
+
+    await page.click('[data-testid="preview-content"]');
+    await page.waitForSelector('[data-testid="hover-card"]', {
+      hidden: true,
+    });
+    await page.waitForSelector('[data-testid="smart-links-preview-modal"]');
+
+    const image = await takeSnapshot(page, 800, 0);
+    expect(image).toMatchProdImageSnapshot();
   });
 });