+++ editor/editor-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-common
 
+## 68.0.2
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 68.0.1
 
 ### Patch Changes